pyproject.toml
[project]
name = "atomic-research-mcp"
version = "0.1.0"
description = "Agentic Research using the Atomic Agents Framework."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.12",
    "atomic-agents>=1.0.21",
    "beautifulsoup4>=4.13.3",
    "instructor>=1.7.2",
    "lxml[html-clean]>=5.3.1",
    "markdownify>=0.14.1",
    "mcp[cli]>=1.3.0",
    "openai>=1.63.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "readability-lxml>=0.8.1",
    "requests>=2.32.3",
]

[project.scripts]
atomic-research = "atomic_research_mcp.server:main"
